Break out virtualize functions into separate modules. Aggregate in index.js (default export).
