+++ src/strings.js
@@ -1,0 +1,3 @@
+export default function(html) {
+    return null;
+}
